TUIO library update

+++ ext/tuio/TuioCursor.cpp
@@ -20,11 +20,11 @@
 
 using namespace TUIO;
 
-TuioCursor::TuioCursor (TuioTime ttime, long si, int ci, float xp, float yp):TuioContainer(ttime,si,xp,yp) {
+TuioCursor::TuioCursor (TuioTime ttime, int si, int ci, float xp, float yp):TuioContainer(ttime,si,xp,yp) {
 	cursor_id = ci;
 }
 
-TuioCursor::TuioCursor (long si, int ci, float xp, float yp):TuioContainer(si,xp,yp) {
+TuioCursor::TuioCursor (int si, int ci, float xp, float yp):TuioContainer(si,xp,yp) {
 	cursor_id = ci;
 }
